[refactor/test] rename package in the make test files
