+++ .circleci/validation/make/makebuild_test.go
@@ -1,4 +1,4 @@
-package validation
+package make
 
 import (
 	"fmt"
