+++ .circleci/validation/make/utilities.go
@@ -1,4 +1,4 @@
-package validation
+package make
 
 import (
 	"bytes"
